@@ -10,18 +10,18 @@
 #  authenticated_at :datetime
 #
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :correspondence do
     name { Faker::Name.name }
     email { Faker::Internet.email }
     email_confirmation { email }
-    category 'general_enquiries'
-    topic { (Faker::Hipster.sentence 4, false, 2)[0..59] }
-    message { Faker::Lorem.paragraph(1) }
+    category { 'general_enquiries' }
+    topic { (Faker::Hipster.sentence word_count: 4, supplemental: false, random_words_to_add: 2)[0..59] }
+    message { Faker::Lorem.paragraph(sentence_count: 1) }
     uuid { SecureRandom.uuid }
 
     trait :authenticated do
-      authenticated_at 10.minutes.ago
+      authenticated_at { 10.minutes.ago }
     end
   end
 end
