@@ -8,10 +8,10 @@
 #  updated_at :datetime         not null
 #
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :feedback do
     ease_of_use {  Settings.feedback_options.sample }
     completeness {  Settings.feedback_options.sample }
-    comment { Faker::Lorem.paragraph(1) }
+    comment { Faker::Lorem.paragraph(sentence_count: 1) }
   end
 end
