@@ -11,7 +11,11 @@
 # end
 
 # These inflection rules are supported but not enabled by default:
+# ActiveSupport::Inflector.inflections(:en) do |inflect|
+#   inflect.acronym 'RESTful'
+# end
  ActiveSupport::Inflector.inflections(:en) do |inflect|
-   inflect.uncountable 'correspondence'
-   inflect.uncountable 'feedback'
+    inflect.uncountable 'correspondence'
+    inflect.uncountable 'feedback'
  end
+ 
